Move StoreContainer.ts and UseCaseContainer.ts to container.ts
