+++ src/commands/CreateLocalTables.ts
@@ -1,8 +1,8 @@
-import { container, TYPES } from "../providers/StoreContainer";
+import { container, TYPES } from "../providers/container";
 import { Todos } from "../usecases/stores/Todos";
 
 container
-  .get<Todos>(TYPES.Todos)
+  .get<Todos>(TYPES.STORE_TODOS)
   .createTable(1 /* readCapacity */, 1 /* writeCapacity */)
   .then(() => {
     console.log("Tables has been created");
